done moving components around. Next step is to review and merge the conflicts that I created when merging pr #111
